requirements.txt
asgiref==3.4.1
boto3==1.20.31
botocore==1.23.31
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.13.0
django-environ==0.8.1
django-heroku==0.0.0
django-storages==1.12.3
flake8==3.9.2
flake8-django==1.1.2
gunicorn==20.1.0
jmespath==0.10.0
mccabe==0.6.1
Pillow==9.0.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
whitenoise==5.3.0
